Sometimes it is needed to pass additional information to the inherited FlaskView classes.
As flask-classy and flask-classful are defined, the __init__ call is
intended to being created without arguments, because when instanced it
is created by just a call without parameters.

While this can be overcomed referencing the needed information from
global objects, this practice is not as clean as clearly defining and
passing the information through the constructor.

This commit adds additional init_argument to register method, optional
with default value of None, that allows to pass information in the
__init__ of the registered classes.

The implementation is fully backwards compatible, with an optional
parameter than when not provided will behave just like it before.

This maybe solves the request in teracyhq/flask-classful/issues#84