Add @method decorator and handling
Using the @method decorator, you can specify for which http method your
class method will be called. This is useful if you cannot use the
default_methods on you class because not all class methods implement the
entire list and if you do not want to use the @route decorator, which
supports method definition as well, because you do not want to lose the
flexibility of classful creating the url rules for you.