Fix compatibility with Werkzeug 2.2
werkzeug.routing.parse_rule was considered an internal detail and has
been removed during a refactor. We can now rely on werkzeug.routing.Rule
which has a `arguments` field that contains the information we need.
However, this field is not populated until the Rule has been added to a
Map and this Map has been bound. So, we use a dummy map and bind it to
the empty string just to get the arguments field ready for our use.