Rewrite test_rule_options
This test file mostly tests the same behavior over and over:
that the options given to `FlaskView.register`, such as
`strict_slashes`, are passed along to `Flask.add_url_rule`.

It did this by testing the behavior of the view when calling
its endpoints with or without the trailing slash, expecting it
to return the same response.

However, werkzeug never guaranteed this behavior. The only
documented behavior tunable by this option is:

"If strict_slashes is enabled (the default), visiting a branch URL
without a trailing slash will redirect to the URL with a slash
appended."

So, with the reimplementation of the routing in werkzeug 2.2, this
behavior is lost.

Instead of testing the behavior of the view, let's mock `add_url_rule`
and check that it's called with `strict_slashes` as `False`.